initial refactor on component scoping
