+++ components/ui/Hero.js
@@ -1,0 +1,13 @@
+<Box
+  sx={{
+    paddingTop: 20,
+    border: '1px solid blue'
+  }}
+>
+  <Grid container p={3} alignItems="center" spacing={3}>
+    <Grid item>
+      <Typography variant="h3">Brooklyn Heights Montessori</Typography>
+    </Grid>
+  </Grid>
+  <Divider />
+</Box>
