+++ components/ui/index.js
@@ -1,0 +1,4 @@
+export { default as PageContainer } from './PageContainer'
+export { default as Radio } from './Radio'
+export { default as Checkbox } from './Checkbox'
+export { default as Chip } from './Chip'
