+++ jsconfig.json
@@ -1,0 +1,10 @@
+{
+  "compilerOptions": {
+    "baseUrl": ".",
+    "paths": {
+      "@ui": ["components/ui"],
+      "@components/*": ["components/*"],
+      "@lib/*": ["lib/*"]
+    }
+  }
+}
